experimentation/pyproject.toml
[project]
name = "constraint-solving"
version = "0.1.0"
description = "Experimental setup for the course CS4535."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "altair>=6.0.0",
    "click>=8.3.1",
    "gitpython>=3.1.45",
    "hvplot>=0.12.1",
    "inquirer>=3.4.1",
    "matplotlib>=3.10.8",
    "polars>=1.35.2",
    "psutil>=7.1.3",
    "pyarrow>=22.0.0",
    "rich>=14.2.0",
    "tomli-w>=1.2.0",
]

[project.scripts]
init_experiment = "constraint_solving.commands.init_experiment:run"
execute_experiment = "constraint_solving.commands.execute_experiment:run"

# Project-specific entry points
generate_mzn_benchmark_set = "constraint_solving.commands.generate_mzn_benchmark_set:run"
process = "constraint_solving.commands.process:run"
generate_figures = "constraint_solving.commands.generate_figures:run"

[build-system]
requires = ["uv_build"]
build-backend = "uv_build"
